// exoplanet-catalog/ClientApp/src/app/Models/PlanetConfiguration.ts
export class PlanetConfiguration{

  planetIdentifier?: string;
  typeFlag?: number;
  planetaryMassJpt?: number;
  radiusJpt?: number;
  periodDays?: number;
  semiMajorAxisAU?: number;
  eccentricity?: string;
  periastronDeg?: string;
  longitudeDeg?: string;
  ascendingNodeDeg?: string;
  inclinationDeg?: number;
  surfaceTempK?: string;
  ageGyr?: string;
  discoveryMethod?: string;
  discoveryYear?: number;
  lastUpdated?: string;
  rightAscension?: string;
  declination?: string;
  distFromSunParsec?: string;
  hostStarMassSlrMass?: number;
  hostStarRadiusSlrRad?: number;
  hostStarMetallicity?: number;
  hostStarTempK?: number;
  hostStarAgeGyr?: string;

}